requirements.txt
streamlit==1.55.0
docling==2.19.0
langchain==1.2.13
langchain-core>=1.2.10,<2.0.0
langchain-experimental>=0.3.0,<1.0.0
langchain-text-splitters>=1.0.0,<2.0.0
langchain-community>=0.3.0,<1.0.0
langchain-groq==1.1.2
langchain-huggingface==1.2.1
rapidocr_onnxruntime==1.4.4
python-dotenv==1.2.2
huggingface-hub==0.36.2
sentence-transformers==5.3.0
torch==2.11.0
faiss-cpu==1.13.2
groq==0.37.1
pydantic==2.12.5
numpy==2.4.3
